add display tab feature

+++ models/tab.js
@@ -3,7 +3,7 @@ module.exports = function(sequelize, DataTypes) {
   var Tab = sequelize.define('Tab', {
     name: DataTypes.STRING,
     file : DataTypes.TEXT,
-    note : DataTypes.INTEGER
+    note : DataTypes.DECIMAL(2,1)
   });
  
   return Tab;
